white_agent/requirements.txt
google-adk>=1.0.0
python-dotenv>=1.0.0
litellm>=1.0.0
